+++ demos/speech_recognition_deepspeech_demo/python/asr_utils/profiles.py
